refactor: inline ConversationEventBus, remove conversation_events.py
