+++ src/reachy_mini_conversation_app/console.py
@@ -15,8 +15,10 @@ import sys
 import time
 import asyncio
 import logging
+import threading
 from typing import List, Optional
 from pathlib import Path
+from collections.abc import Callable, AsyncGenerator
 
 from fastrtc import AdditionalOutputs, audio_to_float32
 from scipy.signal import resample
@@ -64,56 +66,24 @@ except Exception:  # pragma: no cover - only loaded when settings_app is used
     BaseModel = object  # type: ignore
     Request = object  # type: ignore
 
-from reachy_mini_conversation_app.conversation_events import ConversationEventBus
-
-
 logger = logging.getLogger(__name__)
 
-# Toggle which web UI is served by the settings app.
-#
-# The legacy UI lives in ``static/`` and is unchanged. A new web UI lives in
-# ``static_v2/`` and adds a personality picker (with avatars) plus a live
-# conversation orb fed by the ``/conversation_events`` SSE endpoint. The
-# ``REACHY_MINI_UI`` environment variable lets users opt into either one
-# without code changes:
-#
-#   unset / "modern"   -> serve ``static_v2/`` (default on this branch)
-#   "legacy"           -> serve ``static/``
-#
-# Defaulting to the modern UI on this branch keeps day-to-day testing easy
-# while the ``legacy`` switch acts as an instant escape hatch if anything
-# regresses.
+# REACHY_MINI_UI env var: "legacy" → static/, anything else → static_v2/ (default).
 _UI_MODE_ENV = "REACHY_MINI_UI"
 _UI_LEGACY_DIR = "static"
 _UI_MODERN_DIR = "static_v2"
 
 
 def _select_ui_dir_name() -> str:
-    """Return the static folder name to serve, based on ``REACHY_MINI_UI``."""
     mode = (os.environ.get(_UI_MODE_ENV) or "").strip().lower()
     return _UI_LEGACY_DIR if mode == "legacy" else _UI_MODERN_DIR
 
 
-# Interval at which the SSE stream emits a comment line to keep the connection
-# alive across intermediate proxies. 15s is well below the typical 60s idle
-# timeout while staying invisible to the client.
-_SSE_KEEPALIVE_INTERVAL_SECONDS = 15.0
+_SSE_KEEPALIVE_INTERVAL_SECONDS = 15.0  # below typical 60s proxy idle timeout
 
 
 def _detach_framework_root_routes(app: "FastAPI") -> None:
-    """Remove ``GET /`` and ``mount('/static')`` routes pre-registered by the framework.
-
-    The Reachy Mini Apps framework wires its own root + ``/static`` mount in
-    ``ReachyMiniApp.__init__`` (see ``reachy_mini/apps/app.py``). It points at
-    the per-instance ``static/`` folder which, for the modern UI, is exactly
-    what we want to override. We can only drop those routes after the
-    framework has run, hence this surgical cleanup right before we register
-    our own.
-
-    Routes are matched conservatively (``path == "/"`` or ``path == "/static"``)
-    so we never strip routes added by other apps or by user customisations
-    targeting different paths.
-    """
+    """Strip framework-registered GET / and /static routes so ours aren't silently shadowed."""
     routes = getattr(app, "router", None)
     routes = getattr(routes, "routes", None) if routes else getattr(app, "routes", None)
     if routes is None:
@@ -121,32 +91,66 @@ def _detach_framework_root_routes(app: "FastAPI") -> None:
     survivors = []
     for route in routes:
         path = getattr(route, "path", None)
-        if path == "/" or path == "/static":
+        if path in ("/", "/static"):
             logger.debug("detaching framework-provided route %r (%s)", path, type(route).__name__)
             continue
         survivors.append(route)
     routes[:] = survivors
 
 
+class ConversationEventBus:
+    """Thread-safe fan-out bus that delivers conversation activity events to SSE subscribers."""
+
+    MAX_QUEUE_SIZE = 64
+
+    def __init__(self) -> None:
+        """Initialize the bus."""
+        # Each entry is (loop, queue); loop captured at subscribe() time for cross-thread delivery.
+        self._subscribers: list[tuple[asyncio.AbstractEventLoop, "asyncio.Queue[str]"]] = []
+        self._lock = threading.Lock()
+
+    def subscribe(self) -> tuple["asyncio.Queue[str]", Callable[[], None]]:
+        """Return a per-subscriber queue and its unsubscribe callback. Must be called from a coroutine."""
+        loop = asyncio.get_running_loop()
+        queue: "asyncio.Queue[str]" = asyncio.Queue(maxsize=self.MAX_QUEUE_SIZE)
+        entry = (loop, queue)
+        with self._lock:
+            self._subscribers.append(entry)
+
+        def unsubscribe() -> None:
+            with self._lock:
+                try:
+                    self._subscribers.remove(entry)
+                except ValueError:
+                    pass
+
+        return queue, unsubscribe
+
+    def publish(self, event: str) -> None:
+        """Broadcast to all subscribers via call_soon_threadsafe. Drops events for full or closed queues."""
+        with self._lock:
+            snapshot = list(self._subscribers)
+        for loop, queue in snapshot:
+            try:
+                loop.call_soon_threadsafe(self._enqueue_safely, queue, event)
+            except RuntimeError:
+                pass  # loop closed; subscriber will clean itself up on disconnect
+
+    @staticmethod
+    def _enqueue_safely(queue: "asyncio.Queue[str]", event: str) -> None:
+        try:
+            queue.put_nowait(event)
+        except asyncio.QueueFull:
+            logger.debug("conversation event dropped (subscriber queue full): %s", event)
+
+
 async def _conversation_events_stream(
     bus: ConversationEventBus,
     request: "Request",
-):
-    """Yield Server-Sent Events for one conversation event subscriber.
-
-    Format:
-        event: activity
-        data: <reason>
-
-    Plus periodic ``: keep-alive`` comment lines so the connection survives
-    long idle periods. The stream terminates cleanly when the client closes
-    the connection (detected via ``Request.is_disconnected``).
-    """
+) -> "AsyncGenerator[str, None]":
+    """Yield SSE lines for one subscriber; emits keep-alive comments during idle periods."""
     queue, unsubscribe = bus.subscribe()
     try:
-        # Hint the browser to retry quickly on transient drops, and tell the
-        # client we are alive right now so it can flip to the "connected"
-        # state without waiting for the first real event.
         yield "retry: 2000\n\n"
         yield "event: ready\ndata: connected\n\n"
 
@@ -161,6 +165,7 @@ async def _conversation_events_stream(
             yield f"event: activity\ndata: {event}\n\n"
     finally:
         unsubscribe()
+
 
 LOCAL_PLAYER_BACKEND = (
     getattr(MediaBackend, "LOCAL", None)
@@ -222,20 +227,11 @@ class LocalStream:
         self._asyncio_loop = None
         self._active_backend_name = get_backend_choice()
 
-        # Bus that fans out conversation activity events to SSE subscribers
-        # (consumed by the modern web UI orb). The bus is permanently attached
-        # to the handler so it works regardless of which UI is selected; the
-        # legacy UI simply ignores the ``/conversation_events`` endpoint.
         self._event_bus = ConversationEventBus()
         self._attach_event_bus_to_handler()
 
     def _attach_event_bus_to_handler(self) -> None:
-        """Wire the event bus into the realtime handler if it supports it.
-
-        Older or alternative handlers may not implement ``set_activity_observer``
-        (it is opt-in on ``BaseRealtimeHandler``). In that case the bus stays
-        live but receives no events, and SSE clients just see keep-alives.
-        """
+        """Wire the event bus as the handler's activity observer, if supported."""
         setter = getattr(self.handler, "set_activity_observer", None)
         if callable(setter):
             setter(self._event_bus.publish)
@@ -471,21 +467,11 @@ class LocalStream:
         index_file = static_dir / "index.html"
         logger.info("Serving settings UI from %s", static_dir)
 
-        # The Reachy Mini Apps framework (see ``reachy_mini.apps.app``) eagerly
-        # registers ``GET /`` and ``mount("/static", ...)`` against the per-app
-        # ``static/`` folder before this method runs. FastAPI dispatches on the
-        # first matching route, so without the cleanup below our routes would
-        # be silently shadowed and the user would still see the legacy UI even
-        # though our log says we are serving ``static_v2``. We strip those
-        # framework-registered routes first, then enregister ours pointing at
-        # the bundle selected via ``REACHY_MINI_UI``.
+        # Framework pre-registers GET / and /static; strip them so our routes aren't shadowed.
         _detach_framework_root_routes(self._settings_app)
 
         if hasattr(self._settings_app, "mount"):
             try:
-                # Serve /static/* assets from the selected UI folder. The mount
-                # path stays ``/static`` regardless of which folder is active so
-                # the HTML can use stable absolute URLs.
                 self._settings_app.mount("/static", StaticFiles(directory=str(static_dir)), name="static")
             except Exception:
                 pass
@@ -562,12 +548,6 @@ class LocalStream:
                 ready = False
             return JSONResponse({"ready": ready})
 
-        # GET /conversation_events -> Server-Sent Events stream of conversation
-        # activity transitions (user_speech_started, response_created,
-        # assistant_audio_delta, ...). Consumed by the modern web UI orb.
-        # The stream emits raw activity reasons; the client maps them onto its
-        # own visual states. This keeps the backend a dumb event tap and lets
-        # the UI evolve its mapping without backend changes.
         event_bus = self._event_bus
 
         @self._settings_app.get("/conversation_events")
